Save argument counting results to file system, parse raw text

+++ clear-wsd-experiments/src/main/java/edu/colorado/clear/wsd/app/VerbSenseArgumentCounter.java
@@ -17,9 +17,11 @@ import org.mapdb.Serializer;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.ObjectInputStream;
+import java.io.PrintWriter;
 import java.io.Serializable;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -37,8 +39,10 @@ import java.util.stream.Collectors;
 import edu.colorado.clear.wsd.WordSenseAnnotator;
 import edu.colorado.clear.wsd.WordSenseClassifier;
 import edu.colorado.clear.wsd.corpus.CorpusReader;
+import edu.colorado.clear.wsd.corpus.TextCorpusReader;
 import edu.colorado.clear.wsd.corpus.semlink.VerbNetReader;
 import edu.colorado.clear.wsd.feature.annotator.Annotator;
+import edu.colorado.clear.wsd.parser.StanfordDependencyParser;
 import edu.colorado.clear.wsd.type.DepNode;
 import edu.colorado.clear.wsd.type.DependencyTree;
 import edu.colorado.clear.wsd.type.FeatureType;
@@ -104,38 +108,48 @@ public class VerbSenseArgumentCounter {
     }
 
     private static final String COUNTS = "counts";
-    private static final String FILES = "files";
+    private static final String COMPLETE = "complete";
+    private static final String PARSED = "parsed";
 
     @Parameter(names = {"-corpus", "-c"}, description = "Path to corpus directory or file for argument counting", required = true)
     private String corpusPath;
-    @Parameter(names = "-ext", description = "Extension of files in corpus to be processed")
+    @Parameter(names = "-ext", description = "Extension of parsed files in corpus to be processed")
     private String corpusExt = ".dep";
-    @Parameter(names = {"-model", "-m"}, description = "Path to word sense classifier model")
+    @Parameter(names = "-raw", description = "Extension of raw files in corpus to be parsed")
+    private String rawExt = ".txt";
+    @Parameter(names = {"-model", "-m"}, description = "Path to word sense classifier model file")
     private String modelPath;
     @Parameter(names = "-db", description = "Path to MapDB DB file to persist or restore counts", required = true)
     private String dbPath;
     @Parameter(names = {"-outputDir", "o", "-out"}, description = "Path to output directory")
     private String outputPath;
+    @Parameter(names = "-outputExt", description = "Output file extension")
+    private String outputExt = ".txt";
     @Parameter(names = "-relations", description = "Relations to be included in counts")
     private Set<String> relations = Sets.newHashSet("dobj");
     @Parameter(names = "--update", description = "Update existing DB")
     private boolean update = true;
     @Parameter(names = "--overwrite", description = "Overwrite existing DB")
     private boolean overwrite = false;
+    @Parameter(names = "--parseOnly", description = "Only parse, don't count or apply annotator")
+    private boolean parseOnly = false;
 
-    @Parameter(names = "-limit", description = "Maximum number of entries to return in output")
-    private int limit = 1000000;
-    @Parameter(names = "-sort", description = "Ordered list of fields to sort by in output")
+    @Parameter(names = "-limit", description = "Maximum number of entries to return in output", order = 900)
+    private int limit = (int) 10000000;
+    @Parameter(names = "-sort", description = "List of fields to sort by in output", order = 901)
     private List<ArgField> sortFields = Collections.singletonList(ArgField.COUNT);
-    @Parameter(names = "-group", description = "Field to group by in output")
-    private ArgField groupBy = ArgField.SENSE;
-    @Parameter(names = "-print", description = "Ordered list of fields to print in output")
-    private List<ArgField> printFields = Arrays.asList(ArgField.PREDICATE, ArgField.ARGUMENT, ArgField.COUNT);
+    @Parameter(names = "-print", description = "List of fields to print (in order) in output", order = 902)
+    private List<ArgField> printFields = Arrays.asList(ArgField.ARGUMENT, ArgField.COUNT);
+    @Parameter(names = "-group", description = "Field to group by in output", order = 903)
+    private List<ArgField> groupBy = Arrays.asList(ArgField.PREDICATE, ArgField.SENSE);
+    @Parameter(names = {"-separator", "-sep"}, description = "Separator character", order = 904)
+    private String sep = ",";
 
     private CorpusReader<DependencyTree> corpusReader = new VerbNetReader.VerbNetCoNllDepReader();
     private Annotator<DependencyTree> senseAnnotator;
     private DB db;
     private HTreeMap<Argument, Long> countMap;
+    private Set<String> parsed;
     private Set<String> processed;
 
     private VerbSenseArgumentCounter(String... args) {
@@ -154,6 +168,21 @@ public class VerbSenseArgumentCounter {
         }
         initializeDb();
         initializeAnnotator();
+        validateDir();
+    }
+
+    private void validateDir() {
+        if (outputPath == null) {
+            return;
+        }
+        File outputFile = new File(outputPath);
+        if (!outputFile.exists() && !outputFile.mkdirs()) {
+            throw new RuntimeException("Unable to create save directory");
+        }
+        if (outputFile.isFile()) {
+            throw new IllegalArgumentException("Output path provided is an existing file (" + outputPath + "), "
+                    + "please provide a directory path.");
+        }
     }
 
     private void initializeDb() {
@@ -161,26 +190,31 @@ public class VerbSenseArgumentCounter {
             throw new RuntimeException("DB already exists at specified location. Please specify '--update' or '--overwrite' "
                     + "to loading existing DB or overwrite. Otherwise, change the path name.");
         }
+        if (overwrite && new File(dbPath).exists()) {
+            boolean deleted = new File(dbPath).delete();
+            if (!deleted) {
+                throw new RuntimeException("Unable to remove old DB");
+            }
+        }
         db = DBMaker.fileDB(dbPath)
                 .fileMmapEnable()
                 .closeOnJvmShutdown()
                 .make();
         if (overwrite) {
-            boolean deleted = new File(dbPath).delete();
-            if (!deleted) {
-                throw new RuntimeException("Unable to remove old DB");
-            }
             //noinspection unchecked
             countMap = db.<Argument, Long>hashMap(COUNTS, Serializer.JAVA, Serializer.LONG).create();
-            processed = db.hashSet(FILES).serializer(Serializer.STRING).create();
+            parsed = db.hashSet(PARSED).serializer(Serializer.STRING).create();
+            processed = db.hashSet(COMPLETE).serializer(Serializer.STRING).create();
+        } else {
+            //noinspection unchecked
+            countMap = db.<Argument, Long>hashMap(COUNTS, Serializer.JAVA, Serializer.LONG).createOrOpen();
+            parsed = db.hashSet(PARSED).serializer(Serializer.STRING).createOrOpen();
+            processed = db.hashSet(COMPLETE).serializer(Serializer.STRING).createOrOpen();
         }
-        //noinspection unchecked
-        countMap = db.<Argument, Long>hashMap(COUNTS, Serializer.JAVA, Serializer.LONG).createOrOpen();
-        processed = db.hashSet(FILES).serializer(Serializer.STRING).createOrOpen();
     }
 
     private void initializeAnnotator() {
-        if (modelPath != null) {
+        if (modelPath != null && !parseOnly) {
             Stopwatch sw = Stopwatch.createStarted();
             log.debug("Loading sense annotator at {}...", modelPath);
             try (ObjectInputStream ois = new ObjectInputStream(new FileInputStream(modelPath))) {
@@ -196,13 +230,13 @@ public class VerbSenseArgumentCounter {
         }
     }
 
-    private List<File> getCorpusFiles() {
+    private List<File> getCorpusFiles(String ext) {
         List<File> results = new ArrayList<>();
         try {
             File corpusFile = new File(corpusPath);
             if (corpusFile.isDirectory()) {
-                Files.walk(Paths.get(corpusPath))
-                        .filter(path -> path.endsWith(corpusExt))
+                Files.walk(Paths.get(corpusPath), 999)
+                        .filter(path -> path.toString().endsWith(ext))
                         .map(Path::toFile)
                         .forEach(results::add);
             } else {
@@ -215,18 +249,46 @@ public class VerbSenseArgumentCounter {
     }
 
     private void run() {
-        for (File file : getCorpusFiles()) {
-            if (processed.contains(file.getPath())) { // skip already-processed files
-                continue;
-            }
-            List<DependencyTree> instances = readTrees(file);
-            if (senseAnnotator != null) {
-                instances.parallelStream().forEach(senseAnnotator::annotate);
-            }
-            instances.forEach(this::process);
-            processed.add(file.getPath());
+        List<File> toParse = getCorpusFiles(rawExt).stream()
+                .filter(f -> !processed.contains(f.getPath() + corpusExt))
+                .collect(Collectors.toList());
+        if (toParse.size() > 0) {
+            TextCorpusReader reader = new TextCorpusReader(new StanfordDependencyParser());
+            log.debug("Found {} files ending in {} at {}", toParse.size(), rawExt, corpusPath);
+            toParse.parallelStream().forEach(
+                    file -> {
+                        if (!parsed.contains(file.getPath())) {
+                            log.debug("Parsing file {}", file);
+                            try (FileInputStream inputStream = new FileInputStream(file)) {
+                                // read instances
+                                List<DependencyTree> dependencyTrees = reader.readInstances(inputStream);
+                                // write instances
+                                try (FileOutputStream outputStream = new FileOutputStream(file.getPath() + corpusExt)) {
+                                    corpusReader.writeInstances(dependencyTrees, outputStream);
+                                }
+                                parsed.add(file.getPath());
+                            } catch (IOException e) {
+                                log.warn("Unable to process file {}", file, e);
+                            }
+                        }
+                    }
+            );
         }
-        outputEntries();
+        if (!parseOnly) {
+            for (File file : getCorpusFiles(corpusExt)) {
+                if (processed.contains(file.getPath())) { // skip already-processed files
+                    continue;
+                }
+                log.debug("Processing file: {}", file.getAbsolutePath());
+                List<DependencyTree> instances = readTrees(file);
+                if (senseAnnotator != null) {
+                    instances.parallelStream().forEach(senseAnnotator::annotate);
+                }
+                instances.forEach(this::process);
+                processed.add(file.getPath());
+            }
+            outputEntries();
+        }
         db.close();
     }
 
@@ -250,7 +312,7 @@ public class VerbSenseArgumentCounter {
             if (null == sense || sense.isEmpty()) {
                 continue;
             }
-            String lemma = depNode.feature(FeatureType.Lemma);
+            String lemma = depNode.feature(FeatureType.Predicate);
             //noinspection SuspiciousMethodCalls
             depNode.children().stream()
                     .filter(child -> relations.contains(child.feature(FeatureType.Dep)))
@@ -265,9 +327,17 @@ public class VerbSenseArgumentCounter {
     }
 
     private String formatted(Map.Entry<Argument, Long> entry, List<ArgField> fields) {
-        return String.join("\t", fields.stream()
+        return String.join(sep, fields.stream()
                 .map(f -> f.getToString().apply(entry))
                 .collect(Collectors.toList()));
+    }
+
+    private List<ArgumentGroup> group(List<ArgField> fields, List<Map.Entry<Argument, Long>> counts) {
+        ListMultimap<String, Map.Entry<Argument, Long>> index = Multimaps.index(counts, entry ->
+                fields.stream().map(f -> f.getToString().apply(entry)).collect(Collectors.joining(sep)));
+        return index.asMap().entrySet().stream()
+                .map(e -> new ArgumentGroup(e.getKey(), (List<Map.Entry<Argument, Long>>) e.getValue()))
+                .collect(Collectors.toList());
     }
 
     private void outputEntries() {
@@ -276,9 +346,13 @@ public class VerbSenseArgumentCounter {
                 .limit(limit)
                 .collect(Collectors.toList());
         // group entries by provided field
-        List<ArgumentGroup> groups = groupBy == null
+        List<ArgumentGroup> groups = groupBy.size() == 0
                 ? Collections.singletonList(new ArgumentGroup(COUNTS, counts)) : group(groupBy, counts);
         groups.sort(Comparator.comparingLong(ArgumentGroup::getTotal).reversed());
+
+        if (outputPath != null) {
+            log.debug("Saving results to {}", outputPath);
+        }
         for (ArgumentGroup argumentGroup : groups) {
             // sort grouped entries by sort fields
             Comparator<Map.Entry<Argument, Long>> comparator = Comparators.chain(sortFields.stream()
@@ -288,19 +362,23 @@ public class VerbSenseArgumentCounter {
                     .sorted(comparator)
                     .collect(Collectors.toList());
             // print entries with provided formatter
-            System.out.println(argumentGroup.getKey()); // TODO: provide options to output to files
-            for (Map.Entry<Argument, Long> entry : sorted) {
-                System.out.println(formatted(entry, printFields));
+            if (outputPath != null) {
+                try (PrintWriter writer = new PrintWriter(new File(outputPath, argumentGroup.getKey().replaceAll(
+                        "[^a-zA-Z0-9.\\-]", "_") + outputExt))) {
+                    for (Map.Entry<Argument, Long> entry : sorted) {
+                        writer.println(formatted(entry, printFields));
+                    }
+                } catch (FileNotFoundException e) {
+                    log.warn("Unable to write results to file", e);
+                }
+            } else {
+                System.out.println(argumentGroup.getKey());
+                for (Map.Entry<Argument, Long> entry : sorted) {
+                    System.out.println(formatted(entry, printFields));
+                }
+                System.out.println();
             }
-            System.out.println();
         }
-    }
-
-    private List<ArgumentGroup> group(ArgField field, List<Map.Entry<Argument, Long>> counts) {
-        ListMultimap<String, Map.Entry<Argument, Long>> index = Multimaps.index(counts, entry -> field.getToString().apply(entry));
-        return index.asMap().entrySet().stream()
-                .map(e -> new ArgumentGroup(e.getKey(), (List<Map.Entry<Argument, Long>>) e.getValue()))
-                .collect(Collectors.toList());
     }
 
     public static void main(String... args) {
